feat: complete link skill DI
